Vue/Vue&DRF/雛形/requirements.txt
Django~=2.2.10
djangorestframework~=3.11.0 